was bad choice

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -10,7 +10,7 @@
  */
 void print_number(int n)
 {
-	long m, c, num;
+	int m, c, num;
 
 	m = 1;
 	c = 1;
